requirements/local.txt
-r base.txt

django-extensions==2.1.6
factory-boy==2.11.1
pytest-django==3.4.8
pytest-cov==2.6.1
